fix: keep question mapping layer-neutral and route session listing via repository
Addresses architectural review:
- Move PluginQuestionInfo -> shared QuestionInfo mapping into the layer-neutral
  plugin_to_shared_mapping.dart so the SSE BridgeEventMapper (Layer 4) no longer
  imports a repositories/mappers (Layer 2) extension; the repository question
  mapper reuses it from there.
- Add OpenCodeRepository.listSessions and have ActiveSessionTracker.coldStart go
  through it instead of reaching the API directly.
